+++ libs/cardano-ledger-binary/testlib/Test/Cardano/Ledger/Binary/Golden.hs
@@ -10,24 +10,36 @@ module Test.Cardano.Ledger.Binary.Golden (
   expectDecoderFailureAnn,
   expectDecoderResultOn,
   toPackageGolden,
+  goldenForToCBOR,
+  goldenForEncCBOR,
+  cborGoldenSpec,
+  cborAnnGoldenSpec,
 ) where
 
 import Cardano.Ledger.Binary (
   Annotator,
   DecCBOR (..),
   DecoderError,
+  EncCBOR (..),
   ToCBOR (..),
   Version,
   decodeFullAnnotator,
+  serialize,
   toLazyByteString,
  )
 import qualified Cardano.Ledger.Binary as Binary
+import qualified Cardano.Ledger.Binary.Plain as Plain
+import qualified Data.ByteString.Lazy as BSL
 import Data.TreeDiff (ToExpr (..))
-import Data.Typeable (Proxy (..))
-import GHC.Stack (HasCallStack)
+import Data.Typeable (Proxy (..), typeOf)
 import Test.Cardano.Ledger.Binary.Plain.Golden (Enc)
-import Test.Cardano.Ledger.Binary.RoundTrip (embedTripAnnExpectation)
-import Test.Hspec (Expectation, expectationFailure, shouldBe)
+import Test.Cardano.Ledger.Binary.RoundTrip (
+  embedTripAnnExpectation,
+  roundTripAnnRangeExpectation,
+  roundTripCborRangeExpectation,
+ )
+import Test.Cardano.Ledger.Binary.TreeDiff (CBORBytes (..))
+import Test.Hspec
 import qualified Test.Hspec.Golden as Golden (Golden (..))
 
 decodeEnc :: forall a. DecCBOR (Annotator a) => Version -> Enc -> Either DecoderError a
@@ -100,3 +112,73 @@ toPackageGolden mkFullPath g = do
       { Golden.goldenFile = fullPathGoldenFile
       , Golden.actualFile = fullPathActualFile
       }
+
+-- | `Golden` specification for `ToCBOR`
+goldenForToCBOR ::
+  ToCBOR a =>
+  -- | Path to the golden file relative to the package
+  FilePath ->
+  a ->
+  -- | Value, when encoded will be expected to produce same golden file.
+  Golden.Golden BSL.ByteString
+goldenForToCBOR goldenFileName t =
+  Golden.Golden
+    { Golden.output = Plain.serialize t
+    , Golden.encodePretty = show . CBORBytes . BSL.toStrict
+    , Golden.writeToFile = BSL.writeFile
+    , Golden.readFromFile = BSL.readFile
+    , Golden.goldenFile = goldenFileName
+    , Golden.actualFile = Nothing
+    , Golden.failFirstTime = False
+    }
+
+-- | `Golden` specification for `EncCBOR`
+goldenForEncCBOR ::
+  EncCBOR a =>
+  -- | Path to the golden file relative to the package
+  FilePath ->
+  Version ->
+  a ->
+  -- | Value, when encoded will be expected to produce same golden file.
+  Golden.Golden BSL.ByteString
+goldenForEncCBOR goldenFileName version t =
+  Golden.Golden
+    { Golden.output = serialize version t
+    , Golden.encodePretty = show . CBORBytes . BSL.toStrict
+    , Golden.writeToFile = BSL.writeFile
+    , Golden.readFromFile = BSL.readFile
+    , Golden.goldenFile = goldenFileName
+    , Golden.actualFile = Nothing
+    , Golden.failFirstTime = False
+    }
+
+-- | Check `EncCBOR` golden spec as well as roundtripping of the golden example with `DecCBOR`
+cborGoldenSpec ::
+  forall a.
+  ( Eq a
+  , Show a
+  , EncCBOR a
+  , DecCBOR a
+  , HasCallStack
+  ) =>
+  (FilePath -> IO FilePath) -> FilePath -> Version -> a -> Spec
+cborGoldenSpec mkFullPath goldenFileName version a = do
+  describe (show (typeOf a)) $ do
+    it "Golden" $ toPackageGolden mkFullPath $ goldenForEncCBOR goldenFileName version a
+    it "RoundTrip Golden Example" $ roundTripCborRangeExpectation version version a
+
+-- | Check `ToCBOR` golden spec as well as roundtripping of the golden example with `DecCBOR` for
+-- its `Annotator` version.
+cborAnnGoldenSpec ::
+  forall a.
+  ( Eq a
+  , Show a
+  , ToCBOR a
+  , DecCBOR (Annotator a)
+  , HasCallStack
+  ) =>
+  (FilePath -> IO FilePath) -> FilePath -> Version -> a -> Spec
+cborAnnGoldenSpec mkFullPath goldenFileName version a = do
+  describe (show (typeOf a)) $ do
+    it "Golden" $ toPackageGolden mkFullPath $ goldenForToCBOR goldenFileName a
+    it "RoundTrip Golden Example" $ roundTripAnnRangeExpectation version version a