+++ libs/cardano-ledger-binary/CHANGELOG.md
@@ -9,6 +9,7 @@
 
 ### `testlib`
 
+* Add `goldenForEncCBOR` and `cborGoldenSpec`
 * Add `toPackageGolden`
 * Remove `roundTripFailureExpectation`, `roundTripCborExpectation`, and `cborTrip`
 * Add `noTwiddle`, `HuddleEnv`, `toGenEnv`